Use return values for Day 1
I shouldn't go back and modify previous solutions but I had to.

Going forward I will try not to. For instance, I think going forward I'm going to read the input data in main.go and pass the data to the SolveDay... functions. But I won't, or will try not to, go back and change Days 1 and 2.